refactor(docs): circular-progress dx
