+++ apps/docs/content/components/circular-progress/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {CircularProgress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <CircularProgress color="default" aria-label="Loading..."/>
-      <CircularProgress color="primary" aria-label="Loading..."/>
-      <CircularProgress color="secondary" aria-label="Loading..."/>
-      <CircularProgress color="success" aria-label="Loading..."/>
-      <CircularProgress color="warning" aria-label="Loading..."/>
-      <CircularProgress color="danger" aria-label="Loading..."/>
-    </div> 
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
