+++ apps/docs/content/components/circular-progress/label.ts
@@ -1,10 +1,4 @@
-const App = `import {CircularProgress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <CircularProgress label="Loading..." />
-  );
-}`;
+import App from "./label.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
