+++ apps/docs/content/components/circular-progress/sizes.raw.jsx
@@ -1,0 +1,11 @@
+import {CircularProgress} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <CircularProgress aria-label="Loading..." size="sm" />
+      <CircularProgress aria-label="Loading..." size="md" />
+      <CircularProgress aria-label="Loading..." size="lg" />
+    </div>
+  );
+}
