+++ apps/docs/content/components/circular-progress/sizes.ts
@@ -1,14 +1,4 @@
-const App = `import {CircularProgress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <CircularProgress size="sm" aria-label="Loading..."/>
-      <CircularProgress size="md" aria-label="Loading..."/>
-      <CircularProgress size="lg" aria-label="Loading..."/>
-    </div> 
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
