+++ apps/docs/content/components/circular-progress/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {CircularProgress} from "@nextui-org/react";
+
+export default function App() {
+  return <CircularProgress aria-label="Loading..." />;
+}
