+++ apps/docs/content/components/circular-progress/usage.ts
@@ -1,10 +1,4 @@
-const App = `import {CircularProgress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <CircularProgress aria-label="Loading..." />
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
