+++ apps/docs/content/components/circular-progress/value-formatting.raw.jsx
@@ -1,0 +1,14 @@
+import {CircularProgress} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <CircularProgress
+      color="success"
+      formatOptions={{style: "unit", unit: "kilometer"}}
+      label="Speed"
+      showValueLabel={true}
+      size="lg"
+      value={70}
+    />
+  );
+}
