+++ apps/docs/content/components/circular-progress/value-formatting.ts
@@ -1,17 +1,4 @@
-const App = `import {CircularProgress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <CircularProgress
-      label="Speed"
-      size="lg"
-      value={70}
-      color="success"
-      formatOptions={{ style: "unit", unit: "kilometer" }}
-      showValueLabel={true}
-    />
-  );
-}`;
+import App from "./value-formatting.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
