+++ apps/docs/content/components/circular-progress/value.raw.jsx
@@ -1,0 +1,23 @@
+import {CircularProgress} from "@nextui-org/react";
+
+export default function App() {
+  const [value, setValue] = React.useState(0);
+
+  React.useEffect(() => {
+    const interval = setInterval(() => {
+      setValue((v) => (v >= 100 ? 0 : v + 10));
+    }, 500);
+
+    return () => clearInterval(interval);
+  }, []);
+
+  return (
+    <CircularProgress
+      aria-label="Loading..."
+      color="warning"
+      showValueLabel={true}
+      size="lg"
+      value={value}
+    />
+  );
+}
