+++ apps/docs/content/components/circular-progress/value.ts
@@ -1,26 +1,4 @@
-const App = `import {CircularProgress} from "@nextui-org/react";
-
-export default function App() {
-  const [value, setValue] = React.useState(0);
-
-  React.useEffect(() => {
-    const interval = setInterval(() => {
-      setValue((v) => (v >= 100 ? 0 : v + 10));
-    }, 500);
-
-    return () => clearInterval(interval);
-  }, []);
-
-  return (
-    <CircularProgress
-      aria-label="Loading..."
-      size="lg"
-      value={value}
-      color="warning"
-      showValueLabel={true}
-    />
-  );
-}`;
+import App from "./value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
